INSPECT: goodbye ErrorAnnotator
