+++ src/main/java/org/aya/intellij/inspection/AyaInspection.kt
@@ -6,6 +6,11 @@ import com.intellij.codeInspection.ProblemsHolder
 import com.intellij.psi.PsiElementVisitor
 import com.intellij.util.IncorrectOperationException
 import org.aya.intellij.AyaBundle.message
+import org.aya.intellij.inspection.error.ErrorInspection
+import org.aya.intellij.inspection.warning.BadCounterexampleInspection
+import org.aya.intellij.inspection.warning.BadModifierInspection
+import org.aya.intellij.inspection.warning.DominationInspection
+import org.aya.intellij.inspection.warning.NamingInspection
 import org.aya.intellij.lsp.AyaLsp
 
 abstract class AyaInspection : LocalInspectionTool() {
@@ -23,10 +28,13 @@ abstract class AyaInspection : LocalInspectionTool() {
 
   class Provider : InspectionToolProvider {
     override fun getInspectionClasses(): Array<Class<out AyaInspection>> = arrayOf(
+      // warnings
       DominationInspection::class.java,
       NamingInspection::class.java,
       BadModifierInspection::class.java,
       BadCounterexampleInspection::class.java,
+      // errors
+      ErrorInspection::class.java,
     )
   }
 }
