+++ src/main/java/org/aya/intellij/inspection/warning/BadModifierInspection.kt
@@ -1,25 +1,26 @@
-package org.aya.intellij.inspection
+package org.aya.intellij.inspection.warning
 
 import com.intellij.codeInspection.*
 import com.intellij.openapi.project.Project
-import org.aya.concrete.error.BadCounterexampleWarn
+import org.aya.concrete.error.BadModifierWarn
 import org.aya.intellij.AyaBundle
+import org.aya.intellij.inspection.AyaInspection
 import org.aya.intellij.lsp.AyaLsp
-import org.aya.intellij.psi.concrete.AyaPsiOpenKw
+import org.aya.intellij.psi.concrete.AyaPsiFnModifiers
 import org.aya.intellij.psi.concrete.AyaPsiVisitor
 
-class BadCounterexampleInspection : AyaInspection() {
-  override fun getDisplayName() = AyaBundle.message("aya.insp.bad.counter")
+class BadModifierInspection : AyaInspection() {
+  override fun getDisplayName() = AyaBundle.message("aya.insp.bad.modifier")
 
   override fun buildVisitor(lsp: AyaLsp, holder: ProblemsHolder, isOnTheFly: Boolean) = object : AyaPsiVisitor() {
-    override fun visitOpenKw(op: AyaPsiOpenKw) = lsp.warningsAt(op, BadCounterexampleWarn::class.java).forEach { _ ->
+    override fun visitFnModifiers(mod: AyaPsiFnModifiers) = lsp.warningsAt(mod, BadModifierWarn::class.java).forEach { _ ->
       holder.registerProblem(
         holder.manager.createProblemDescriptor(
-          op, op,
-          AyaBundle.message("aya.insp.bad.counter"),
+          mod, mod,
+          AyaBundle.message("aya.insp.bad.modifier.info", mod.text),
           ProblemHighlightType.LIKE_DEPRECATED, isOnTheFly,
           object : LocalQuickFix {
-            override fun getFamilyName() = CommonQuickFixBundle.message("fix.remove", op.text)
+            override fun getFamilyName() = CommonQuickFixBundle.message("fix.remove", mod.text)
             override fun applyFix(project: Project, descriptor: ProblemDescriptor) {
               descriptor.psiElement.prevSibling.delete() // the whitespace
               descriptor.psiElement.delete()
