+++ src/main/java/org/aya/intellij/inspection/warning/DominationInspection.kt
@@ -1,8 +1,9 @@
-package org.aya.intellij.inspection
+package org.aya.intellij.inspection.warning
 
 import com.intellij.codeInspection.*
 import com.intellij.openapi.project.Project
 import org.aya.intellij.AyaBundle
+import org.aya.intellij.inspection.AyaInspection
 import org.aya.intellij.lsp.AyaLsp
 import org.aya.intellij.psi.AyaPsiElement
 import org.aya.intellij.psi.concrete.AyaPsiBareClause
