+++ src/main/resources/aya/aya-bundle.properties
@@ -1,4 +1,5 @@
 aya.group.name=Aya prover
+aya.insp.error=Error
 aya.insp.dom=Unreachable clause can be removed
 aya.insp.shadow=The name shadows a previous local definition from outer scope
 aya.insp.bad.modifier=Meaningless modifier
