+++ src/main/resources/inspectionDescriptions/Error.html
@@ -1,0 +1,5 @@
+<html>
+<body>
+This inspection shows errors reported by Aya compiler.
+</body>
+</html>
